+++ wrappers/go/foundation/base64.go
@@ -13,10 +13,10 @@ type Base64 struct {
 /*
 * Calculate length in bytes required to hold an encoded base64 string.
 */
-func Base64EncodedLen(dataLen int) int {
+func Base64EncodedLen(dataLen uint) uint {
     proxyResult := /*pr4*/C.vscf_base64_encoded_len((C.size_t)(dataLen)/*pa10*/)
 
-    return int(proxyResult) /* r9 */
+    return uint(proxyResult) /* r9 */
 }
 
 /*
@@ -24,7 +24,7 @@ func Base64EncodedLen(dataLen int) int {
 * Note, written buffer is NOT null-terminated.
 */
 func Base64Encode(data []byte) []byte {
-    strBuf, strBufErr := bufferNewBuffer(int(Base64EncodedLen(len(data)) /* lg1 */))
+    strBuf, strBufErr := bufferNewBuffer(int(Base64EncodedLen(uint(len(data))) /* lg1 */))
     if strBufErr != nil {
         return nil
     }
@@ -39,17 +39,17 @@ func Base64Encode(data []byte) []byte {
 /*
 * Calculate length in bytes required to hold a decoded base64 string.
 */
-func Base64DecodedLen(strLen int) int {
+func Base64DecodedLen(strLen uint) uint {
     proxyResult := /*pr4*/C.vscf_base64_decoded_len((C.size_t)(strLen)/*pa10*/)
 
-    return int(proxyResult) /* r9 */
+    return uint(proxyResult) /* r9 */
 }
 
 /*
 * Decode given data from the base64 format.
 */
 func Base64Decode(str []byte) ([]byte, error) {
-    dataBuf, dataBufErr := bufferNewBuffer(int(Base64DecodedLen(len(str)) /* lg1 */))
+    dataBuf, dataBufErr := bufferNewBuffer(int(Base64DecodedLen(uint(len(str))) /* lg1 */))
     if dataBufErr != nil {
         return nil, dataBufErr
     }
